add User Management, Slider and Cards at Home View

+++ hogwartsExpress/routes/web.php
@@ -19,18 +19,18 @@ Auth::routes();
 
 // R del CRUD
 Route::get('/', [TrainRouteController::class, 'index'])->name('home');
-Route::get('/home', [TrainRouteControllerr::class, 'index']);
+Route::get('/home', [TrainRouteController::class, 'index']);
 
 //D del CRUD
-Route::delete('/delete/{id}', [TrainRouteController::class, 'destroy'])->name('deleteTrainRoute');
+Route::delete('/delete/{id}', [TrainRouteController::class, 'destroy'])->name('deleteTrainRoute')->middleware('isadmin', 'auth');
 
 //C del CRUD
-Route::get('/create', [TrainRouteController::class, 'create'])->name('createTrainRoute');
-Route::post('/', [TrainRouteController::class, 'store'])->name('storeTrainRoute');
+Route::get('/create', [TrainRouteController::class, 'create'])->name('createTrainRoute')->middleware('isadmin', 'auth');
+Route::post('/', [TrainRouteController::class, 'store'])->name('storeTrainRoute')->middleware('isadmin', 'auth');
 
 //U del CRUD
-Route::get('/edit/{id}', [TrainRouteController::class, 'edit'])->name('editTrainRoute');
-Route::patch('/trainRoute/{id}', [TrainRouteController::class, 'update'])->name('updateTrainRoute');
+Route::get('/edit/{id}', [TrainRouteController::class, 'edit'])->name('editTrainRoute')->middleware('isadmin', 'auth');
+Route::patch('/trainRoute/{id}', [TrainRouteController::class, 'update'])->name('updateTrainRoute')->middleware('isadmin', 'auth');
 
 //Show
 Route::get('/show/{id}', [TrainRouteController::class, 'show'])->name('showTrainRoute');
